requirements.txt
numpy
mypy
pillow~=11.3.0
opencv-python
matplotlib